Support deferred chunk reset in command buffers
This change allows resetting CmdAllocator without norifying command
buffers that use it, assuming that those command buffers are not going
to be used between the allocator reset and CB Begin/Reset/Free.

In order to handle the reset that has happened, a generation number is
added to CmdAllocator. It is incremented each reset and CmdBuffers and
CmdStream use it to detect that CmdAllocator was reset. If it was, the
users simply clear their chunk lists as opposed to resetting and
returning the chunks.